+++ src/main/java/fr/patedor/PFR_Equipe/repository/EmployeRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface EmployeRepository extends JpaRepository<Employe, Integer> {
 
@@ -15,4 +16,6 @@ public interface EmployeRepository extends JpaRepository<Employe, Integer> {
     Employe findByPrenom(String prenom);
 
     Employe findByNom(String nom);
+
+    Optional<Employe> findByLogin(String login);
 }
