+++ src/main/java/fr/patedor/PFR_Equipe/security/jwt/AuthenticationService.java
@@ -1,7 +1,8 @@
 package fr.patedor.PFR_Equipe.security.jwt;
 
+import fr.patedor.PFR_Equipe.entity.Employe;
 import fr.patedor.PFR_Equipe.entity.Utilisateur;
-import fr.patedor.PFR_Equipe.repository.UtilisateurRepository;
+import fr.patedor.PFR_Equipe.repository.EmployeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -12,7 +13,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class AuthenticationService {
     @Autowired
-    private UtilisateurRepository utilisateurRepository;
+    private EmployeRepository employeRepository;
     @Autowired
     private AuthenticationManager authenticationManager;
     @Autowired
@@ -23,8 +24,8 @@ public class AuthenticationService {
                 new UsernamePasswordAuthenticationToken(request.getLogin(), request.getMdp()));
 
         //Génère le token
-        Utilisateur utilisateur = utilisateurRepository.findByLogin(request.getLogin()).orElseThrow();
-        String jwtToken = jwtService.generateToken(utilisateur);
+        Employe employe = employeRepository.findByLogin(request.getLogin()).orElseThrow();
+        String jwtToken = jwtService.generateToken(employe);
         AuthenticationResponse authResponse = new AuthenticationResponse();
         authResponse.setToken(jwtToken);
