+++ src/main/java/fr/patedor/PFR_Equipe/security/jwt/JwtConfig.java
@@ -1,6 +1,6 @@
 package fr.patedor.PFR_Equipe.security.jwt;
 
-import fr.patedor.PFR_Equipe.repository.UtilisateurRepository;
+import fr.patedor.PFR_Equipe.repository.EmployeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -16,12 +16,12 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 public class JwtConfig {
 
     @Autowired
-    private UtilisateurRepository userRepository;
+    private EmployeRepository employeRepository;
 
     //Instancie un userDetailsService, qui est juste un DAO qui a accès uniquement à findByLogin
     @Bean
     UserDetailsService userDetailsService() {
-        return login -> userRepository.findByLogin(login)
+        return login -> employeRepository.findByLogin(login)
                 .orElseThrow(() -> new UsernameNotFoundException("User not found"));
     }
