+++ src/main/resources/application.properties
@@ -5,8 +5,8 @@ app.jwt.secret=40f12e49b6e34d121a67ee7b00d4dfe6e3237e3ac057b5280cf4a3238c9c0946b
 # Configuration de la datasource (SQL Server)
 spring.datasource.driver-class-name=com.microsoft.sqlserver.jdbc.SQLServerDriver
 spring.datasource.url=jdbc:sqlserver://localhost;databaseName=PFR;trustServerCertificate=true
-spring.datasource.username=purukogi
-spring.datasource.password=purukogi
+spring.datasource.username=Purukogi
+spring.datasource.password=PurukogiDB
 
 # Hibernate (Gestion ORM)
 spring.jpa.hibernate.ddl-auto=none
